requirements.txt
yara-python==3.11.0
requests==2.22.0
bs4==0.0.1
rich==11.0.0
